# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateRoleRequest:
    properties:
      is_admin_group:
        type: boolean
      is_protected:
        type: boolean
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      is_admin_group:
        type: boolean
      is_protected:
        type: boolean
      name:
        maxLength: 50
        minLength: 3
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  models.UpdateRoleRequest:
    properties:
      is_admin_group:
        type: boolean
      is_protected:
        type: boolean
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: WattWatch API server.
  title: WattWatch API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: List all roles with optional filtering
      parameters:
      - description: Search term for role name
        in: query
        name: search
        type: string
      - description: Filter by protected status
        in: query
        name: protected
        type: boolean
      - description: Filter by admin group status
        in: query
        name: admin_group
        type: boolean
      - description: Field to order by
        in: query
        name: order_by
        type: string
      - description: Order descending
        in: query
        name: order_desc
        type: boolean
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new role (admin only)
      parameters:
      - description: Role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - admin only
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Role already exists
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role (admin only)
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - admin only
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Role in use by users
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Get a role by its ID (admin only)
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - admin only
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update a role's details (admin only)
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid request body or role ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - admin only
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Role already exists
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - roles
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"